requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
cachetools==5.5.2y
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.1
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
frozenlist==1.7.0
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-dialogflow==2.41.1
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.4.4
phonenumbers==9.0.6
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyngrok==7.2.9
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
retell-sdk==4.32.0
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
twilio==9.6.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
whitenoise==6.9.0
yarl==1.20.1
